+++ v3/crates/compatibility/src/compatibility_date.rs
@@ -182,5 +182,7 @@ pub fn get_compatibility_date_for_flag(flag: Flag) -> Option<CompatibilityDate> 
             Some(new_compatibility_date(2025, 4, 24))
         }
         Flag::SendMissingArgumentsToNdcAsNulls => Some(new_compatibility_date(2025, 5, 30)),
+        Flag::DisallowLiteralsAsBooleanExpressionArguments => None, // todo add docs PR and update
+                                                                    // before release
     }
 }